==> processing/otros/05-long.R <==
# Libraries
if (! require("pacman")) install.packages("pacman")

pacman::p_load(tidyverse,
               car,
               sjmisc, 
               here,
               sjlabelled,
               SciViews,
               naniar)


options(scipen=999)
rm(list = ls())

# Data

load(url("https://dataverse.harvard.edu/api/access/datafile/10797987"))

glimpse(elsoc_long_2016_2023)


# Processing

elsoc_long_2016_2023[elsoc_long_2016_2023 ==-999] <- NA
elsoc_long_2016_2023[elsoc_long_2016_2023 ==-888] <- NA
elsoc_long_2016_2023[elsoc_long_2016_2023 ==-777] <- NA
elsoc_long_2016_2023[elsoc_long_2016_2023 ==-666] <- NA

elsoc_long_2016_2023 <- elsoc_long_2016_2023 %>% 
  mutate(just_educ = d02_02, 
         just_pension = d02_01, 
         just_health = d02_03, 
         merit_effort = c18_09,
         merit_talent = c18_10, 
         age = m0_edad, m01, 
         sex = m0_sexo, 
         ess_d = d01_01, 
         ess_o = d01_02,
         ideo = c15,
         ) %>% 
  as_tibble() %>% 
  sjlabelled::drop_labels(., drop.na = FALSE)

# Market Justice Preferences

frq(elsoc_long_2016_2023$just_health)
frq(elsoc_long_2016_2023$just_pension)
frq(elsoc_long_2016_2023$just_educ)

elsoc_long_2016_2023 <- cbind(elsoc_long_2016_2023, "mjp" = rowMeans(elsoc_long_2016_2023 %>% select(just_health, just_pension, just_educ), na.rm=TRUE))

summary(elsoc_long_2016_2023$mjp)

elsoc_long_2016_2023 <- elsoc_long_2016_2023 %>% 
  mutate(
    across(
      .cols = c(just_health, just_pension, just_educ),
      .fns = ~ car::recode(., recodes = c("1='Strongly disagree'; 2='Disagree';
                                          3='Neither agree nor disagree'; 4='Agree';
                                          5='Strongly agree'"), 
                           levels = c("Strongly disagree", "Disagree", "Neither agree nor disagree", "Agree", "Strongly agree"),
                           as.factor = T)
    )
  )

elsoc_long_2016_2023$just_health <- sjlabelled::set_label(elsoc_long_2016_2023$just_health, 
                                                          label = "Health distributive justice")

elsoc_long_2016_2023$just_pension <- sjlabelled::set_label(elsoc_long_2016_2023$just_pension, 
                                                           label = "Pension distributive justice")

elsoc_long_2016_2023$just_educ <- sjlabelled::set_label(elsoc_long_2016_2023$just_educ, 
                                                        label = "Education distributive justice")

elsoc_long_2016_2023$mjp <- sjlabelled::set_label(elsoc_long_2016_2023$mjp, 
                                                  label = "Market justice preferences")

# ess
sjmisc::frq(elsoc_long_2016_2023$ess_d)
sjmisc::frq(elsoc_long_2016_2023$ess_o)

elsoc_long_2016_2023$ess_d <- as.numeric(elsoc_long_2016_2023$ess_d)
elsoc_long_2016_2023$ess_o <- as.numeric(elsoc_long_2016_2023$ess_o)

elsoc_long_2016_2023$mobility <- (elsoc_long_2016_2023$ess_d - elsoc_long_2016_2023$ess_o)

sjmisc::frq(elsoc_long_2016_2023$mobility)

elsoc_long_2016_2023 <- elsoc_long_2016_2023 %>% 
  rowwise() %>% 
  mutate(mobility_f = case_when(mobility <= -1 ~ "Donward",
                                mobility == 0 ~ "Non-mobile",
                                mobility >= 1 ~ "Upward",
                                TRUE ~ NA_character_),
         mobility_f = factor(mobility_f,
                             levels = c("Non-mobile", "Donward", "Upward"))) %>% 
  ungroup()

frq(elsoc_long_2016_2023$mobility_f)

# Meritocracy

frq(elsoc_long_2016_2023$merit_effort)
frq(elsoc_long_2016_2023$merit_talent)

elsoc_long_2016_2023 <- elsoc_long_2016_2023 %>% 
  mutate(
    across(
      .cols = c(merit_effort, merit_talent),
      .fns = ~ car::recode(., recodes = c("1='Strongly disagree'; 2='Disagree';
                                          3='Neither agree nor disagree'; 4='Agree';
                                          5='Strongly agree'"), 
                           levels = c("Strongly disagree", "Disagree", "Neither agree nor disagree", "Agree", "Strongly agree"),
                           as.factor = T)
    )
  )


elsoc_long_2016_2023$merit_effort <- sjlabelled::set_label(elsoc_long_2016_2023$merit_effort, 
                                                           label = "People are rewarded for their efforts")

elsoc_long_2016_2023$merit_talent <- sjlabelled::set_label(elsoc_long_2016_2023$merit_talent, 
                                                           label = "People are rewarded for their intelligence")

# Controls

# sex
elsoc_long_2016_2023$sex <- car::recode(elsoc_long_2016_2023$sex, 
                                        recodes = c("1='Male'; 2='Female'"), 
                                        levels = c("Male", "Female"),
                                        as.factor = T)

elsoc_long_2016_2023$sex <- sjlabelled::set_label(elsoc_long_2016_2023$sex, 
                                                  label = "Gender")

# age
frq(elsoc_long_2016_2023$age)

elsoc_long_2016_2023$age <- 
  factor(car::recode(elsoc_long_2016_2023$age, 
                     "18:29=1;30:49=2;50:64=3;65:150=4"),
         labels = c('18-29', '30-49', '50-64', '65 or more'))
elsoc_long_2016_2023$age <-
  sjlabelled::set_label(elsoc_long_2016_2023$age, 
                        label = c("Age groups")) 


# political indentification

frq(elsoc_long_2016_2023$ideo)

elsoc_long_2016_2023$ideo<-
  factor(
    car::recode(
      elsoc_long_2016_2023$ideo,
      "c(11,12,-888,-999)='Does not identify';c(0,1,2,3,4)='Left';
     c(5)='Center';c(6,7,8,9,10)='Right'"
    ),
    levels = c('Left', 'Center', 'Right', 'Does not identify')
  )

elsoc_long_2016_2023$ideo<- factor(elsoc_long_2016_2023$ideo,levels = levels(elsoc_long_2016_2023$ideo))

elsoc_long_2016_2023$ideo <- 
  sjlabelled::set_label(x = elsoc_long_2016_2023$ideo, 
                        label = "Political identification") 

frq(elsoc_long_2016_2023$ideo)


# Socioeconomic_________________________________________________________________

# Education_______________________________________
elsoc_long_2016_2023$educ <- 
  car::recode(elsoc_long_2016_2023$m01,
              "c(1,2,3,4,5,6,7)=1;c(8,9,10)=2; c(-888,-999)=NA")
elsoc_long_2016_2023$educ <-
  factor(elsoc_long_2016_2023$educ,
         labels = c("Less than Universitary","Universitary"))

#reverse education, reference level is the highest level
#elsoc_long_2016_2023$educ <- forcats::fct_rev(elsoc_long_2016_2023$educ)

elsoc_long_2016_2023$educ <- 
  sjlabelled::set_label(x = elsoc_long_2016_2023$educ,
                        label = "Education")
sjmisc::frq(elsoc_long_2016_2023$educ)

#Recoding of education to years based on casen 2017.
elsoc_long_2016_2023$educyear<- as.numeric(
  car::recode(elsoc_long_2016_2023$m01, 
              "1=0;2=4.3;3=7.5;4=9.8;5=12.02;6=13.9;
               7=14.8;8=14.9;9=16.9;10=19.07;c(-888,-999)=NA", 
              as.numeric = T))

elsoc_long_2016_2023$educyear <- 
  sjlabelled::set_label(x = elsoc_long_2016_2023$educyear,
                        label = "Education in years")

class(elsoc_long_2016_2023$educyear)
sjmisc::frq(elsoc_long_2016_2023$educyear)

# N Household:
# Select variables______________________________________________________________

df_study1 <- 
  elsoc_long_2016_2023 %>% 
  filter(muestra == 1) %>% #keep original sample (vs refresh)
  select(idencuesta,ola,region,region_cod,comuna,comunacod=comuna_cod,
         muestra,
         ponderador_long_total, segmento, estrato,
         just_educ,
         just_pension,
         just_health,
         mjp,
         merit_effort,
         merit_talent,
         mobility,
         mobility_f,
         educ,
         educyear,
         sex,
         age,
         ideo,
         nhogar1,
         m46_nhogar,
         m54, m30, m30b, m29)

# Household income_________________________________________

#Impute midpoint of income ranges
df_study1$m30_rec <-
  as.numeric(car::recode(df_study1$m30,
                         "1=110000;2=251000;3=305000;4=355000;5=400000;
            6=445000;7=490000;8=535000;9=585000;10=640000;11=700000;12=765000;
            13=845000;14=935000;15=1040000;16=1180000;17=1375000;18=1670000;
            19=2275000;20=2700000;NA=NA;c(-888,-999)=NA"))

#Impute midpoint of income ranges (2021)
df_study1$m30b_rec <-
  as.numeric(car::recode(df_study1$m30b,
                         "1=125000;2=300000;3=400000;4=575000;5=70000;NA=NA;c(-888,-999)=NA"))

sjmisc::frq(df_study1$m30b_rec)

#Recode DK/DA of Income to NA
df_study1$m29_rec <-
  as.numeric(car::recode(df_study1$m29,"c(-888,-999)=NA"))

#replace NA of income with new imputed variable
df_study1$m29_imp <- 
  ifelse(test = !is.na(df_study1$m29_rec),
         yes =  df_study1$m29_rec,
         no =  df_study1$m30_rec)
summary(df_study1$m29_imp)

df_study1$m29_imp <- 
  ifelse(test = is.na(df_study1$m29_imp),
         yes =  df_study1$m30b_rec,
         no =  df_study1$m29_imp)
summary(df_study1$m29_imp)

# deflate at each year's prices
df_study1$deflactor <-
  with(df_study1, case_when(
    ola == 2016 ~ 113.88 / 123.82,
    ola == 2017 ~ 116.46 / 123.82,
    ola == 2018 ~ 119.45 / 123.82,
    ola == 2019 ~ 123.82 / 123.82
  ))

# Reshape long to wide
df_study1_wide <- df_study1 %>% 
  tidyr::pivot_wider(id_cols = c("idencuesta","muestra"),
                     names_from = "ola",
                     # names_prefix = c("educ","educyear","ing_pc","quintil","quintil1"),
                     values_from = names(select(df_study1,comuna,comunacod,ponderador_long_total,segmento, estrato, just_educ:ideo, nhogar1, m46_nhogar, m54, m30, m30b, m29_imp))
  )

df_study1_wide$m54_6 <- df_study1_wide$m54_7

# reshape from long to wide

df_study1_long <- df_study1_wide %>%
  pivot_longer(
    cols = -c(idencuesta, muestra),
    names_to = c(".value", "ola"),
    names_pattern = "^(comuna|comunacod|ponderador_long_total|segmento|estrato|just_educ|just_pension|just_health|mjp|merit_effort|merit_talent|mobility|mobility_f|educ|educyear|sex|age|ideo|nhogar1|m46_nhogar|m54|m30|m30b|m29_imp)_(\\d+)$"
  ) %>%
  mutate(ola = as.integer(ola))

df_study1_long <-
  df_study1_long %>%
  mutate(n_hogar =
           dplyr::case_when(ola == 1 ~ nhogar1,
                            ola == 2 ~ m46_nhogar,
                            ola == 3 ~ m54,
                            ola == 4 ~ m54,
                            ola == 5 ~ m54,
                            ola == 6 ~ m54,
                            ola == 7 ~ m54))
sjmisc::frq(df_study1_long$n_hogar)

#Recode DK/DA to NA
df_study1_long$n_hogar_r<-
  car::recode(df_study1_long$n_hogar,"c(-888,-999)=NA")

# Per capita household income:
df_study1_long$ing_pc <- 
  (df_study1_long$m29_imp/df_study1_long$n_hogar_r)

df_study1_long$ing_pc <-
  sjlabelled::set_label(x = df_study1_long$ing_pc,
                        label = "Household income per capita")  

sjmisc::descr(df_study1_long$ing_pc)

# Compute income quintiles
df_study1_long <- df_study1_long %>% 
  group_by(ola) %>% 
  mutate(quintil = ntile(-desc(ing_pc), 5)) %>% 
  ungroup()

df_study1_long$quintil <- 
  factor(df_study1_long$quintil,
         levels = c(1, 2, 3, 4, 5),
         labels = c('Q1', 'Q2', 'Q3', 'Q4', 'Q5')) # Quintiles as factors

#reverse quintile, reference level is the highest quintile
#elsoc_long_2016_2023$quintil <- forcats::fct_rev(elsoc_long_2016_2023$quintil)

df_study1_long$quintil <- 
  sjlabelled::set_label(x = df_study1_long$quintil,
                        label = "Household income quintile per capita")  

sjmisc::frq(df_study1_long$quintil)

#include new quintile category with missing cases
df_study1_long$quintil1<-
  car::recode(df_study1_long$quintil, 
              "'Q1'='Q1';'Q2'= 'Q2';'Q3'='Q3';'Q4'='Q4';'Q5'='Q5'; NA='QNA'")

#elsoc_long_2016_2023$quintil1 <- factor(elsoc_long_2016_2023$quintil1, c("Q1","Q2","Q3","Q4","Q5","QNA"))

df_study1_long$quintil1 <- 
  sjlabelled::set_label(x = df_study1_long$quintil1,
                        label = "Household income quintile per capita (NA)") 
sjmisc::frq(df_study1_long$quintil1)

# Reshape long to wide
df_study1_wide <- df_study1_long %>% 
  tidyr::pivot_wider(id_cols = c("idencuesta","muestra"),
                     names_from = "ola",
                     # names_prefix = c("educ","educyear","ing_pc","quintil","quintil1"),
                     values_from = names(select(df_study1_long,comuna,comunacod,ponderador_long_total,segmento, estrato, just_educ:ideo, quintil, quintil1, ing_pc)))

# fix data to w01 values
df_study1_wide$age_2 <-df_study1_wide$age_1 #age
df_study1_wide$age_3 <-df_study1_wide$age_1
df_study1_wide$age_4 <-df_study1_wide$age_1
df_study1_wide$age_5 <-df_study1_wide$age_1
df_study1_wide$age_6 <-df_study1_wide$age_1
df_study1_wide$age_7 <-df_study1_wide$age_1

df_study1_wide$sex_2 <-df_study1_wide$sex_1 #sex
df_study1_wide$sex_3 <-df_study1_wide$sex_1
df_study1_wide$sex_4 <-df_study1_wide$sex_1
df_study1_wide$sex_5 <-df_study1_wide$sex_1
df_study1_wide$sex_6 <-df_study1_wide$sex_1
df_study1_wide$sex_7 <-df_study1_wide$sex_1

df_study1_wide$educ_2 <-df_study1_wide$educ_1 #education
df_study1_wide$educ_3 <-df_study1_wide$educ_1
df_study1_wide$educ_4 <-df_study1_wide$educ_1
df_study1_wide$educ_5 <-df_study1_wide$educ_1
df_study1_wide$educ_6 <-df_study1_wide$educ_1
df_study1_wide$educ_7 <-df_study1_wide$educ_1

df_study1_wide$educyear_2 <-df_study1_wide$educyear_1 #education years
df_study1_wide$educyear_3 <-df_study1_wide$educyear_1
df_study1_wide$educyear_4 <-df_study1_wide$educyear_1
df_study1_wide$educyear_5 <-df_study1_wide$educyear_1
df_study1_wide$educyear_6 <-df_study1_wide$educyear_1
df_study1_wide$educyear_7 <-df_study1_wide$educyear_1

df_study1_wide$quintil1_2 <-df_study1_wide$quintil1_1 #quintiles
df_study1_wide$quintil1_3 <-df_study1_wide$quintil1_1
df_study1_wide$quintil1_4 <-df_study1_wide$quintil1_1
df_study1_wide$quintil1_5 <-df_study1_wide$quintil1_1
df_study1_wide$quintil1_6 <-df_study1_wide$quintil1_1
df_study1_wide$quintil1_7 <-df_study1_wide$quintil1_1

df_study1_wide$ing_pc_2 <-df_study1_wide$ing_pc_1 #income per capita
df_study1_wide$ing_pc_3 <-df_study1_wide$ing_pc_1
df_study1_wide$ing_pc_4 <-df_study1_wide$ing_pc_1
df_study1_wide$ing_pc_5 <-df_study1_wide$ing_pc_1
df_study1_wide$ing_pc_6 <-df_study1_wide$ing_pc_1
df_study1_wide$ing_pc_7 <-df_study1_wide$ing_pc_1

df_study1_wide$ideo_2 <-df_study1_wide$ideo_1 # political position
df_study1_wide$ideo_3 <-df_study1_wide$ideo_1
df_study1_wide$ideo_4 <-df_study1_wide$ideo_1
df_study1_wide$ideo_5 <-df_study1_wide$ideo_1
df_study1_wide$ideo_6 <-df_study1_wide$ideo_1
df_study1_wide$ideo_7 <-df_study1_wide$ideo_1


dim(df_study1_wide)

sjPlot::view_df(df_study1_wide,
                show.frq = T,show.values = T,show.na = T,show.prc = T, show.type = T)

# reshape from long to wide
df_study1_long <- df_study1_wide %>%
  pivot_longer(
    cols = -c(idencuesta, muestra),
    names_to   = c(".value", "ola"),
    names_pattern = paste0(
      "^(comuna|comunacod|ponderador_long_total|segmento|estrato|just_educ|",
      "just_pension|just_health|mjp|merit_effort|merit_talent|mobility|",
      "mobility_f|educ|educyear|sex|age|ideo|quintil|",
      "quintil1|ing_pc)_(\\d+)$"
    )
  ) %>%
  mutate(ola = as.integer(ola))

names(df_study1_long) #check names of long dataset
dim(df_study1_long) #check dimensions of the dataframe
# Original dataset with 7 waves
df_study2_long <- df_study1_long

# filter the dataset for the waves 1 to 4 and 6 to 7
df_study1_long <-
  df_study1_long %>% 
  filter(ola %in% c(1,2,3,4,6,7)) %>% 
  mutate(ola=factor(ola,levels = 1:7,labels = 1:7))
dim(df_study1_long) #check, now is OK

# df_study1_long <- 
# set_label(x = df_study1_long,
#           label = get_label(select(df_study1,names(df_study1_long))))


sjPlot::view_df(df_study1_long,
                show.frq = T,show.values = T,show.na = T,show.prc = T,
                show.type = T)
#______________________________________________________________________________
# obtain the idencuesta for wave 7
ids <- 
  df_study1 %>% 
  select(idencuesta,ola) %>% 
  filter(ola==7) %>% 
  sjmisc::frq(idencuesta,show.na = F) %>% as.data.frame()


# filter data by the idencuesta of t7
df_study1_long_t7 <- 
  df_study1_long %>%
  filter(idencuesta %in% ids$val)

names(df_study1_long_t7)
dim(df_study1_long_t7)
sjmisc::frq(df_study1_long_t7$ola)

save(df_study1_long,file = here::here("input/data/proc/df_study1_long.RData"))
save(df_study1_long_t7,file = here::here("input/data/proc/df_study1_long_t7.RData"))
save(df_study2_long,file = here::here("input/data/proc/df_study2_long.RData"))
